refactor: :recycle: review markups
